+++ CHANGES.md
@@ -6,6 +6,7 @@ See also https://pvp.haskell.org/faq
    https://github.com/haskell-unordered-containers/hashable/pull/263
  * Drop support for GHCs prior GHC-8.2
    The recent `unordered-containers` releases support only GHC-8.2+
+ * Add instance for `OsString`, `PosixString`, `WindowsString` from `filepath-1.4.100.1`
 
 ## Version 1.4.1.0